% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInput.R
\name{dataInput}
\alias{dataInput}
\title{Convenience Function to Prepare Required Data Inputs to \code{IntHTEcf()}}
\usage{
dataInput(data, outcome.model, ps.model)
}
\arguments{
\item{data}{A data.frame object. Must contain the outcome of interest,
the treatment and all model covariates.}

\item{outcome.model}{A formula object. Must be of the form LHS ~ RHS. The
LHS is the outcome of interest. The RHS defines the main effects model and
the Z covariates of the treatment effect function. For example, for outcome,
Y, treatment A, and covariates X1, X2, \code{outcome.model = Y ~ X1 + X2*A}
sets the design matrix of the main effects model as (1, X1, X2) and the
Z of the HTE function as (1, X2). Note that smoothing splines will
automatically be applied for regression method "gam" and should not
be specified here.}

\item{ps.model}{A formula object. Must be of the form LHS ~ RHS. The
LHS is the treatment variable. The RHS defines the propensity score model.
Note that smoothing splines will
automatically be applied for regression method "gam" and should not
be specified here.}
}
\value{
A list object appropriately formatted for use in \code{IntHTEcf()}.
\itemize{
\item \code{X} is a numeric covariate matrix without an intercept. It must contain
all covariates required of the main effects, contrasts, and propensity
score models.
\item \code{Y} is a binary or continuous response vector.
\item \code{A} is a binary treatment vector.
\item \code{mainName} is an integer or a character vector. If a character vector,
the column headers of X defining the main effects component of the
outcome model. Note that an intercept is always assumed and should not be
provided. An intercept only model can be specified as \code{mainName = 1L}.
\item \code{contName} is an integer or a character vector. If a character vector,
the column headers of X defining Z used to estimate the treatment effect.
Note that an intercept is always assumed and should not be
provided. An intercept only model can be specified as \code{contName = 1L}.
\item \code{psName} is an integer or a character vector. If a character vector,
the column headers of X defining the propensity score model.
Note that an intercept is always assumed and should not be
provided. An intercept only model can be specified as \code{psName = 1L}.
}
}
\description{
Provided a data.frame and models for the outcome and propensity regressions,
generates an list of the format required by \code{IntHTEcf()}.
}
