% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toyData.R
\name{elasticToy.bin}
\alias{elasticToy.bin}
\alias{elasticToy.bin.rct}
\alias{elasticToy.bin.rwe}
\title{Toy Binary Outcome Dataset}
\format{
elasticToy.bin provides two datasets. The elasticToy.bin.rct
100 participant records; elasticToy.bin.rwe 500 participant records. Each
list provides the following:
\describe{
\item{Y}{A binary outcome.}
\item{X}{A matrix of two covariates, X1, and X2}
\item{A}{A binary treatment variable}
}
}
\usage{
data("elasticToy.bin")
}
\description{
These datasets are provided only to facilitate examples. They are not based
on or representative of any real-world applications.
}
\keyword{datasets}
