% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInput.R
\name{dataInput}
\alias{dataInput}
\title{Prepare Required Data Inputs to \code{elasticHTE()}}
\usage{
dataInput(data, outcome.model, ps.model)
}
\arguments{
\item{data}{A data.frame object. Must contain the outcome of interest,
the treatment and all model covariates.}

\item{outcome.model}{A formula object. Must be of the form LHS ~ RHS. The
LHS is the outcome of interest. The RHS defines the main effects model and
the Z covariates of the treatment effect function; i.e., of the form
LHS ~ main_effects + treatment_variable * Z.}

\item{ps.model}{A formula object. Must be of the form LHS ~ RHS. The
LHS is the treatment variable. The RHS defines the propensity score model.}
}
\value{
A list object appropriately formatted for use in \code{elasticHTE()}.
\itemize{
\item \code{X}: The covariate matrix excluding an intercept. It
contains all covariates required of the outcome and propensity
score models.
\item \code{Y} The response vector.
\item \code{A} The treatment vector.
\item \code{mainName} The column headers of X defining the main effects
component of the outcome model or 1, indicating an intercept-only model.
\item \code{contName} The column headers of X defining Z or 1, indicating
an intercept-only model.
\item \code{psName} The column headers of X defining the propensity score
model or 1, indicating an intercept-only model.
}
}
\description{
Provided a data.frame and models for the outcome and propensity regressions,
generates an list of the format required by \code{elasticHTE()}.
}
\examples{
data("elasticToy.cont")

data_object <- dataInput(elasticToy.cont.rct,
                         outcome.model = Y ~ X1 + X2 * A,
                         ps.model = A ~ X1 * X2)

}
