% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticHTE.R, R/s3_methods.R
\name{elasticHTE}
\alias{elasticHTE}
\alias{print.elasticHTE}
\alias{summary.elasticHTE}
\title{Elastic Integrative Analysis for Heterogeneous Treatment Effect}
\usage{
elasticHTE(
  data.rct,
  data.rwe,
  mainName = NULL,
  contName = NULL,
  psName = NULL,
  thres.psi = NULL,
  sieve.degree = 2L,
  outcome.type = c("cont", "bin"),
  outcome.method = c("glm", "sl"),
  outcome.controls = list(family = "gaussian"),
  ps.method = c("glm", "sl"),
  ps.controls = list(family = "quasibinomial"),
  fixed = FALSE,
  n.pert = 100L,
  n.boot = 100L,
  n.gamma = 1000L
)

\method{print}{elasticHTE}(x, ...)

\method{summary}{elasticHTE}(object, ...)
}
\arguments{
\item{data.rct}{A named list containing elements "X", "Y", and "A" for an RCT.
Element X is a numeric covariate matrix without an intercept.
Element Y is a binary or continuous response vector.
Element A is a binary treatment vector.
List can also contain element "ps", a vector of known propensity scores P(A=1).}

\item{data.rwe}{A named list containing elements "X", "Y", and "A" for an RWE
study.
Element X is a numeric covariate matrix without an intercept.
Element Y is a binary or continuous response vector.
Element A is a binary treatment vector.}

\item{mainName}{NULL, character vector, or an integer. The covariates of the
main effects component of the outcome model. If NULL, all covariates in
\code{data.rct$X} specify the model; if a character vector, the column
headers of \code{data.rct$X} to include in the model. Note that an
intercept is always included in the model; though it is not recommended,
an intercept only main effects model can be specified as \code{mainName = 1}.}

\item{contName}{NULL, character vector, or an integer. The covariates of the
contrasts component of the outcome model. If NULL, all covariates in
\code{data.rct$X}  specify the model; if a character vector, the column
headers of \code{data.rct$X}  to include in the model. Note these are the
covariates that interact with the treatment variable (~ A:contName); an
intercept is always included, such that ~A is the minimal contrasts model;
though it is not recommended, an intercept only model can be specified as
\code{contName = 1}.}

\item{psName}{NULL, character vector, or an integer. The covariates of the
the propensity score model. If NULL, all covariates in \code{data.rct$X}
specify the model; if a character vector, the column headers of
\code{data.rct$X} to include in the model. Note that an intercept is
always included in the model; an intercept only model can be specified as
\code{psName = 1}.}

\item{thres.psi}{NULL or a scalar numeric. The threshold for constructing
adaptive confidence interval. STH QUESTION: Is there an allowed range we
should mention?}

\item{sieve.degree}{A positive integer > 1. The order of the polynomial
defining the sieve model. For example, \code{sieve.degree = 2} yields outcome and
propensity models that include all covariates, the squared covariates, and
all pair-wise interactions.}

\item{outcome.type}{A character. The type of outcome. Must be one of
\{"cont", "bin"\} indicating a continuous or binary outcome, respectively.}

\item{outcome.method}{A character. The regression tool to be used for
estimating the parameters of the outcome model. Must be one of
\{"glm", "sl"\} indicating the \code{stats::glm()} function or the
\code{SuperLearner::SuperLearner()} function.}

\item{outcome.controls}{A named list. Additional inputs provided to the
outcome regression tool. Element names must match the formal arguments
of the chosen \code{outcome.method}. Should include, at a minimum, element
"family" indicating the error distribution. Please see documentation
of the selected regression tool for additional input options.
For \code{outcome.method = "glm"}, formal arguments \code{formula},
\code{data}, \code{weights}, and \code{subset} are set internally.
And for \code{outcome.method = "sl"}, formal arguments \code{Y},
\code{X}, \code{newX}, and  \code{obsWeight} are set internally.}

\item{ps.method}{A character. The regression tool to be used for
estimating the parameters of the propensity score model. Must be one of
\{"glm", "sl"\} indicating the \code{stats::glm()} function or the
\code{SuperLearner::SuperLearner()} function.}

\item{ps.controls}{A named list. Additional inputs provided to the
propensity score regression tool. Element names must match the formal
arguments of the chosen \code{ps.method}. Should include, at a minimum, element
"family" indicating the error distribution. Please see documentation
of the selected regression tool for additional input options.
For \code{ps.method = "glm"}, formal arguments \code{formula},
\code{data}, \code{weights}, and \code{subset} are set internally.
And for \code{ps.method = "sl"}, formal arguments \code{Y},
\code{X}, \code{newX}, and  \code{obsWeight} are set internally.}

\item{fixed}{A logical. How to select the tuning parameter
\eqn{c_{\gamma}}{c_gamma}. FALSE, the default, selects an adaptive
selection strategy; TRUE selects a fixed threshold strategy.
The default fixed threshold is \code{stats::qchisq(0.95, df = p)},
in which p is the dimension of the contrasts model.}

\item{n.pert}{An integer. The number of perturbations to generate when
estimating the variance.}

\item{n.boot}{An integer. The number of bootstrap samples to generate
when estimating the confidence intervals.}

\item{n.gamma}{An integer. The number of samples to generate to estimate
\eqn{c_{\gamma}}{c_gamma}, the threshold.}

\item{x}{An object of S3 class \code{elasticHTE}.}

\item{...}{Ignored}

\item{object}{An object of S3 class \code{elasticHTE}.}
}
\value{
A list with components:
\item{psi }{A matrix containing the estimated \eqn{\psi}{`psi`} associated
with the treatment modifiers under various models
(\eqn{\psi_p}{`psi_p`}, \eqn{\psi_{eff}}{`psi_eff`}, \eqn{\psi_{rt}}{`psi_rt`},
\eqn{\psi_{elastic}}{`psi_elastic`},
\eqn{\psi_{elastic.debiased}}{`psi_elastic.debiased`}).}
\item{ve }{A matrix containing the estimated standard error for
\eqn{\psi}{`psi`}.}
\item{CIs.inf, CIs.sup }{A matrix containing the estimated confidence
intervals for \eqn{\psi}{`psi`}.}
\item{conservative }{A logical \eqn{I(Tstat < thres.psi)}{I(Tstat < thres.psi)} }
\item{CI.settings }{A list of the settings used in the confidence interval
procedure.}
\item{nuispar }{A list providing the selected \eqn{\gamma}{gamma} and
its corresponding threshold value \eqn{c_{\gamma}}{c.gamma};
indicator \eqn{I(c_{\gamma}  > Tstat)}{I(c.gamma > Tstat)} and its
p-value as well as a list of the settings used in the
selection procedure.}
}
\description{
\code{elasticHTE()} is a test-based dynamic borrowing framework combining
a randomized clinical trial (RCT) and a real-world evidence (RWE) study,
in which a preliminary test statistic is used to gauge the comparability
and reliability of the RWE and to decide whether or not to use the RWE in
an integrative analysis. The parameter of interest is \eqn{\psi}{`psi`},
which quantifies how the treatment effect varies over the treatment
modifiers.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print(elasticHTE)}: Print summary of analysis

\item \code{summary(elasticHTE)}: Summary of analysis.

}}
\examples{
# load provided illustrative toy dataset with continuous outcome
data("elasticToy.cont")

# conduct the elastic integrative analysis with defaults
result.cont <- elasticHTE(data.rct = elasticToy.cont.rct,
                          data.rwe = elasticToy.cont.rwe)

# conduct the elastic integrative analysis using SuperLearner with
# default controls
\dontrun{
result.cont.sl <- elasticHTE(data.rct = elasticToy.cont.rct,
                             data.rwe = elasticToy.cont.rwe,
                             outcome.method = "sl",
                             outcome.control = list("family" = gaussian(),
                                                    "SL.library" = "SL.glm"),
                             ps.method = "sl",
                             ps.control = list("family" = binomial(),
                                               "SL.library" = "SL.glm"))
}
# load provided illustrative toy dataset with binary outcome
data("elasticToy.bin")

# conduct the elastic integrative analysis with defaults
result.bin <- elasticHTE(data.rct = elasticToy.bin.rct,
                         data.rwe = elasticToy.bin.rwe,
                         outcome.type = "bin",
                         outcome.controls = list("family" = "quasibinomial"))

# conduct the elastic integrative analysis using SuperLearning with
# default controls
\dontrun{
result.bin.sl <- elasticHTE(data.rct = elasticToy.bin.rct,
                            data.rwe = elasticToy.bin.rwe,
                            outcome.type = "bin",
                            outcome.method = "sl",
                            outcome.controls = list("family" = "quasibinomial",
                                               "SL.library" = "SL.glm"),
                            ps.method = "sl",
                            ps.control = list("family" = binomial(),
                                               "SL.library" = "SL.glm"))
}
}
