% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toyData.R
\name{elasticToy.cont}
\alias{elasticToy.cont}
\alias{elasticToy.cont.rct}
\alias{elasticToy.cont.rwe}
\title{Toy Continuous Outcome Dataset}
\format{
elasticToy.cont provides two datasets. The elasticToy.cont.rct
100 participant records; elasticToy.cont.rwe 500 participant records. Each
data.frame provides the following:
\itemize{
\item Y: A continuous outcome.
\item X1: A continuous covariate.
\item X2: A continuous covariate.
\item A: A binary treatment variable
}
}
\usage{
data("elasticToy.cont")
}
\description{
These datasets are provided only to facilitate examples. They are not based
on or representative of any real-world applications.
}
\keyword{datasets}
